requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
asyncpg==0.30.0
pyjwt==2.10.1
bcrypt==4.3.0
httpx==0.28.1
python-dotenv==1.1.0
